moved resample out of parsers
