+++ src/diffpy/utils/resample.py
